+++ guide/CHANGELOG.md
@@ -6,6 +6,20 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 Warning: Before deploying in production a bot that can generate heavy traffic, please contact ALE.
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
+### [2.25.2-lts.2] - 2023-10-XX
+#### Removed
+-   None
+
+#### Fixed
+-   Fix `ContactsService::getAll` to return the Contacts in cache the network of the connected user. If you need the previous behaviour use the new API `ContactsService::getAllContactsInCache`.
+-   Fix `InvitationsService` when an error occurs during call of `ContactsService::getContactById`. 
+
+#### Added
+-   Add `ContactsService::getAllContactsInCache` to get the list of _contacts that are in the cache of the current instance of the connected users..
+
+#### Changed
+-   Update `ContactsService::getAll` API result (see Fixed).
+
 ### [2.25.2-lts.1] - 2023-10-03
 #### Removed
 -   None